Pass port through to userauth.
I'm going to want to use it in an upcoming commit, because together
with 'savedhost', it forms the identification of an SSH server (at
least as far as the host key cache is concerned, and therefore it's
appropriate for other uses too).

We were already passing the hostname through for use in user-facing
prompts (not to mention the FQDN version for use in GSSAPI).