Fix aes-select.c macros again.
I decided that the 'namemaker' system introduced recently in commit
fbb979aa985cdcd was just too marginal to be sensible, and it's easier
to simply quote the full SSH id for each protocol.

Also, included an empty argument at the end of each macro invocation,
so that the variadic "..." is never completely missing.